improve custom ai dialog
